Setup LISP family support, emmet_ls and improve formatting
Install nvim-parinfer and Conjure and integrate it with cmp completion.
Change the localleader to "," so that Conjure related keybindings
doesn't interfere with nvim-tree binding <leader>e.

Set filetypes for which prettier is used and set it to use semicolon in
javascript.

Setup emmet_ls for html, css, js and others and get rid of tailwindcss
LSP.