Setup lsp, null-ls, autopairs and comment.nvim
Mason and mason-lspconfig are always loaded because it is what installs
lsp, while treesitter is loaded when I am sure that I want syntax
highlighting in a file type. Whenever that happens then it also likely
that formatting and linting is also needed. That's why null-ls loads
after treesitter. Although lsp is not loaded always. Since it may take
time, it is only loaded when InsertEnter i.e, along with cmp-nvim-lsp
and vim-illuminate. That's why it is important to keep formatting and
linting seperate from lsp and based solely on null-ls, coz I may need
linting and formatting without entering in insert mode.

Once InsertEnter then nvim-autopairs is also loaded.

Comment.nvim is loaded whenever BufReadPre of BufNewFile