Disable cmp in NvimTree and lazyload cmp-conjure
Cmp will not be turned on when searching for files in nvim-tree.
cmp-conjure will not be added as a source until the filetype is scheme,
and when it is added it will get rid of all other sources for that
file. This could be fixed by using a better lua function that instead of
setting the source as cmp-conjure, it appends it to the list of existing
sources.